chore: archive Nelson build workspace folder (#1431)
This PR adds archiving functionality to the Windows CI/CD pipeline by creating a zip archive of the Nelson build workspace directory. The change enables preservation of the complete workspace state for builds on the main/master branches.

Adds a new step to zip the Nelson workspace directory using 7-Zip
Archives are created only for builds on master or main branches
Excludes the Release directory from the archive to reduce size